templebot/requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
configobj==5.0.6
contextlib2==0.6.0.post1
discord.py==1.3.3
docopt==0.6.2
Flask==1.1.2
Flask-WTF==0.14.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
Hypercorn==0.9.5
hyperframe==5.2.0
idna==2.9
itsdangerous==1.1.0
jikanpy==4.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
motor==2.1.0
multidict==4.7.5
munch==2.5.0
num2words==0.5.10
Pillow==7.1.2
pip-autoremove==0.9.1
priority==1.3.0
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
Quart==0.11.5
requests==2.23.0
schema==0.7.2
secure==0.2.1
six==1.14.0
toml==0.10.0
typing-extensions==3.7.4.2
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
wsproto==0.15.0
WTForms==2.3.1
yarl==1.4.2
